Fix constraints package name typo
